+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For more information about the changes, please refer to the [Bootstrap v5 migrat
 ### Changes
 
 - update `README` to document updated UI bootstrap version
+- updated dependencies
 
 ## v2.0.1 (2025-01-21)